check for pid.txt existence, rename update.sh

+++ stop.sh
@@ -1,1 +1,5 @@
-kill -SIGTERM $(cat pid.txt) && rm pid.txt
+touch pid.txt
+if cat pid.txt | grep ''  > /dev/null
+then
+    kill -SIGTERM $(cat pid.txt) && rm pid.txt
+fi
